nba-stat-projections/backend/requirements.txt
nba_api>=1.3.1
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.1
supabase>=2.3.1
python-dotenv>=1.0.0
httpx>=0.27.0
pandas>=2.2.0
numpy>=1.26.0
pytest>=7.4.0 